hvalfangst_function/requirements.txt
azure-functions==1.7.0
pandas~=2.1.4
scikit-learn~=1.3.2